V3.0

+++ libft/ft_atoi.c
@@ -6,23 +6,23 @@
 /*   By: omeslall <omeslall@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/10 18:51:26 by omeslall          #+#    #+#             */
-/*   Updated: 2021/11/16 14:54:46 by omeslall         ###   ########.fr       */
+/*   Updated: 2022/02/11 18:53:12 by omeslall         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"libft.h"
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int			i;
-	int			count;
-	long int	nbr;
+	long		count;
+	long long	nbr;
 
 	i = 0;
 	count = 1;
 	nbr = 0;
 	while (str[i] == '\n' || str[i] == '\r' || str[i] == '\t'
-		   || str[i] == '\v' || str[i] == '\f' || str[i] == ' ')
+		|| str[i] == '\v' || str[i] == '\f' || str[i] == ' ')
 		i++;
 	if (str[i] == '-')
 		count = -1;
@@ -35,4 +35,3 @@ int	ft_atoi(const char *str)
 	}
 	return (nbr * count);
 }
-
